+++ src/scandeval/model_setups/hf.py
@@ -246,18 +246,6 @@ class HFModelSetup:
         if load_in_4bit and importlib.util.find_spec("bitsandbytes") is None:
             raise NeedsExtraInstalled(extra="generative")
 
-        bnb_config = (
-            BitsAndBytesConfig(
-                load_in_4bit=load_in_4bit,
-                bnb_4bit_compute_dtype=(
-                    torch.bfloat16 if torch.cuda.is_bf16_supported() else torch.float16
-                ),
-                bnb_4bit_use_double_quant=True,
-            )
-            if load_in_4bit
-            else None
-        )
-
         config = self._load_hf_model_config(
             model_id=model_id,
             num_labels=dataset_config.num_labels,
@@ -265,6 +253,21 @@ class HFModelSetup:
             label2id=dataset_config.label2id,
             revision=model_config.revision,
             model_cache_dir=model_config.model_cache_dir,
+        )
+
+        use_bf16 = (
+            self.benchmark_config.device == torch.device("cuda")
+            and torch.cuda.is_bf16_supported()
+            and config.to_dict().get("torch_dtype") == "bfloat16"
+        )
+        bnb_config = (
+            BitsAndBytesConfig(
+                load_in_4bit=load_in_4bit,
+                bnb_4bit_compute_dtype=torch.bfloat16 if use_bf16 else torch.float16,
+                bnb_4bit_use_double_quant=True,
+            )
+            if load_in_4bit
+            else None
         )
 
         use_vllm = (